+++ src/components/ProfileDetailsActionMenu/ProfileDetailsActionMenu.js
@@ -6,6 +6,7 @@ import {
   Button,
   Icon,
 } from '@folio/stripes/components';
+import { useStripes } from '@folio/stripes/core';
 
 export const ProfileDetailsActionMenu = ({
   isProfileUsed,
@@ -21,12 +22,16 @@ export const ProfileDetailsActionMenu = ({
     onToggle();
   };
 
+  const stripes = useStripes();
+  const hasOnlyViewPerms = stripes.hasPerm('settings.data-export.view') && !stripes.hasPerm('settings.data-export.enabled');
+
+
   return (
     <>
       <Button
         data-test-edit-profile-button
         buttonStyle="dropdownItem"
-        disabled={isDefaultProfile || isProfileUsed}
+        disabled={isDefaultProfile || isProfileUsed || hasOnlyViewPerms}
         onClick={buildButtonClickHandler(onEdit)}
       >
         <Icon icon="edit">
@@ -37,6 +42,7 @@ export const ProfileDetailsActionMenu = ({
         data-test-duplicate-profile-button
         buttonStyle="dropdownItem"
         onClick={buildButtonClickHandler(onDuplicate)}
+        disabled={hasOnlyViewPerms}
       >
         <Icon icon="duplicate">
           <FormattedMessage id="stripes-data-transfer-components.duplicate" />
@@ -45,7 +51,7 @@ export const ProfileDetailsActionMenu = ({
       <Button
         data-test-delete-profile-button
         buttonStyle="dropdownItem"
-        disabled={isDefaultProfile || isProfileUsed}
+        disabled={isDefaultProfile || isProfileUsed || hasOnlyViewPerms}
         onClick={buildButtonClickHandler(onDelete)}
       >
         <Icon icon="trash">